+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+### Features
+
+* Default to `date_range` fields for all datetime-related types
+
 ### Bug fixes
 
 * Prevent crashing when an error occurs while loading table data - the error will be propely logged instead
+* Fix errors when attempting to filter on `NaiveDatetime` attribuse
 
 ## v0.2.0 (2025-06-18)